+++ apps/meteor/app/livechat/server/lib/logger.ts
@@ -2,3 +2,4 @@ import { Logger } from '@rocket.chat/logger';
 
 export const callbackLogger = new Logger('[Omnichannel] Callback');
 export const businessHourLogger = new Logger('Business Hour');
+export const livechatLogger = new Logger('Livechat');
